Set Resource to internal trace instrumentation
